split orderstop

+++ MQL4/Order/OrderStopPending.mqh
@@ -1,5 +1,5 @@
 //+------------------------------------------------------------------+
-//|                                              OrderStopBroker.mqh |
+//|                                             OrderStopPending.mqh |
 //|                                                   Enrico Lambino |
 //|                             https://www.mql5.com/en/users/iceron |
 //+------------------------------------------------------------------+
@@ -8,64 +8,96 @@
 //+------------------------------------------------------------------+
 //|                                                                  |
 //+------------------------------------------------------------------+
-class COrderStopBroker : public COrderStopBrokerBase
+class COrderStopPending : public COrderStopPendingBase
   {
 public:
-                     COrderStopBroker(void);
-                    ~COrderStopBroker(void);
+                     COrderStopPending(void);
+                    ~COrderStopPending(void);
    virtual void      Check(double &);
 protected:
    virtual bool      ModifyStops(const double,const double);
    virtual bool      ModifyStopLoss(const double);
    virtual bool      ModifyTakeProfit(const double);
-   virtual bool      UpdateOrderStop(const double,const double);
+   
   };
 //+------------------------------------------------------------------+
 //|                                                                  |
 //+------------------------------------------------------------------+
-COrderStopBroker::COrderStopBroker(void)
+COrderStopPending::COrderStopPending(void)
   {
   }
 //+------------------------------------------------------------------+
 //|                                                                  |
 //+------------------------------------------------------------------+
-COrderStopBroker::~COrderStopBroker(void)
+COrderStopPending::~COrderStopPending(void)
   {
   }
+
 //+------------------------------------------------------------------+
 //|                                                                  |
 //+------------------------------------------------------------------+
-bool COrderStopBroker::UpdateOrderStop(const double stoploss,const double takeprofit)
-  {
-   bool modify_sl=false,modify_tp=false;
-   if(stoploss>0)
-      modify_sl=m_stop.MoveStopLoss(m_order.Ticket(),stoploss);
-   if(takeprofit>0)
-      modify_tp=m_stop.MoveTakeProfit(m_order.Ticket(),takeprofit);
-   return modify_tp||modify_sl;
-  }
-//+------------------------------------------------------------------+
-//|                                                                  |
-//+------------------------------------------------------------------+
-void COrderStopBroker::Check(double &volume)
+void COrderStopPending::Check(double &volume)
   {
    if(!CheckPointer(m_stop) || !Active())
       return;
    if(m_order.IsClosed() || m_order.IsSuspended())
      {
       bool delete_sl=false,delete_tp=false;
-      delete_sl=DeleteStopLoss();
-      delete_tp=DeleteTakeProfit();
+      if(!m_stop.Main())
+        {
+         if(m_stop.DeleteStopOrder(m_stoploss_ticket))
+            delete_sl=DeleteStopLoss();
+         if(m_stop.DeleteStopOrder(m_takeprofit_ticket))
+            delete_tp=DeleteTakeProfit();
+        }
       if(delete_sl && delete_tp)
          DeleteEntry();
+      return;
      }
+   if(m_stop.Main())
+     {
+      if(OrderSelect((int)m_order.Ticket(),SELECT_BY_TICKET))
+        {
+         if(OrderCloseTime()>0)
+           {
+            m_stoploss_closed=true;
+            m_takeprofit_closed=true;
+           }
+        }
+     }
+   if(CheckPointer(m_objsl))
+      m_stoploss_closed=m_stop.CheckStopOrder(volume,m_stoploss_ticket);
+   if(CheckPointer(m_objtp))
+      m_takeprofit_closed=m_stop.CheckStopOrder(volume,m_takeprofit_ticket);
+   if(m_stoploss_closed || m_takeprofit_closed)
+     {
+      if(m_stop.OCO())
+        {
+         if(m_stoploss_closed && !m_takeprofit_closed)
+           {
+            if(m_stop.DeleteStopOrder(m_takeprofit_ticket))
+               m_takeprofit_closed=true;
+           }
+         if(m_takeprofit_closed && !m_stoploss_closed)
+           {
+            if(m_stop.DeleteStopOrder(m_stoploss_ticket))
+               m_stoploss_closed=true;
+           }
+        }
+      if(m_stoploss_closed)
+         DeleteStopLoss();
+      if(m_takeprofit_closed)
+         DeleteTakeProfit();
+     }
+   if(((m_stoploss_closed || !CheckPointer(m_objsl)) && (m_takeprofit_closed || !CheckPointer(m_objtp))) || volume<=0)
+      DeleteStopLines();
   }
 //+------------------------------------------------------------------+
 //|                                                                  |
 //+------------------------------------------------------------------+
-bool COrderStopBroker::ModifyStopLoss(const double stoploss)
+bool COrderStopPending::ModifyStopLoss(const double stoploss)
   {
-   bool modify=m_stop.MoveStopLoss(m_order.Ticket(),stoploss);
+   bool modify=m_stop.OrderModify(m_stoploss_ticket,stoploss);
    if(modify)
       MoveStopLoss(stoploss);
    return modify;
@@ -73,9 +105,9 @@ bool COrderStopBroker::ModifyStopLoss(const double stoploss)
 //+------------------------------------------------------------------+
 //|                                                                  |
 //+------------------------------------------------------------------+
-bool COrderStopBroker::ModifyTakeProfit(const double takeprofit)
+bool COrderStopPending::ModifyTakeProfit(const double takeprofit)
   {
-   bool modify=m_stop.MoveTakeProfit(m_order.Ticket(),takeprofit);
+   bool modify=m_stop.OrderModify(m_takeprofit_ticket,takeprofit);
    if(modify)
       MoveTakeProfit(takeprofit);
    return modify;
@@ -83,14 +115,8 @@ bool COrderStopBroker::ModifyTakeProfit(const double takeprofit)
 //+------------------------------------------------------------------+
 //|                                                                  |
 //+------------------------------------------------------------------+
-bool COrderStopBroker::ModifyStops(const double stoploss,const double takeprofit)
+bool COrderStopPending::ModifyStops(const double stoploss,const double takeprofit)
   {
-   bool modify=m_stop.Move(m_order.Ticket(),stoploss,takeprofit);
-   if(modify)
-   {
-      MoveStopLoss(stoploss);
-      MoveTakeProfit(takeprofit);
-   }   
-   return modify;
+   return ModifyStopLoss(stoploss) && ModifyTakeProfit(takeprofit);
   }
 //+------------------------------------------------------------------+
